Adjust the tutorial instructions for supplied text
Before, the tutorial instructions assumed that the students would be
typing all of its text in by hand.  Which would be terrible.  Instead,
I will be giving them its content as plain unmarked text, and they will
simply mark up and cross-link the text using Sphinx.  So I have
rephrased several parts of the tutorial instructions to reflect this.